% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/url.R
\name{cvap_census_url}
\alias{cvap_census_url}
\title{Get Zip File URL for CVAP Special Tabulation Data}
\usage{
cvap_census_url(year = 2022)
}
\arguments{
\item{year}{numeric. Year for the data in 2009 to 2022.}
}
\value{
string
}
\description{
Get Zip File URL for CVAP Special Tabulation Data
}
\examples{
cvap_census_url()
}
\concept{raw}
