% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw.R
\name{cvap_get_raw}
\alias{cvap_get_raw}
\title{Get Raw Citizen Voting Age Population Files}
\usage{
cvap_get_raw(url, out_dir)
}
\arguments{
\item{url}{URL to CVAP zip to download. Use \code{cvap_census_url()}.}

\item{out_dir}{Directory to unzip to. Defaults to temp directory.}
}
\value{
string, path where the data is saved
}
\description{
Get Raw Citizen Voting Age Population Files
}
\examples{
\donttest{
# takes 10-20 seconds
path <- cvap_get_raw(cvap_census_url())
}

}
\concept{raw}
