% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{cvap_process_file}
\alias{cvap_process_file}
\title{Process Census CVAP File}
\usage{
cvap_process_file(path, year, out_dir, moe = TRUE, csv = FALSE)
}
\arguments{
\item{path}{path to csv file}

\item{year}{file year}

\item{out_dir}{directory to create files in}

\item{moe}{Boolean. Default is TRUE. Should margin of error be kept?}

\item{csv}{Boolean. Default is FALSE, which creates an rds file instead.}
}
\value{
tibble of cvap data
}
\description{
Process Census CVAP File
}
\examples{
path <- fs::path_package('cvap', 'extdata', 'County.csv')
cvap_process_file(path, year = 2019, out_dir = tempdir())
}
\concept{raw}
