% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{de_block}
\alias{de_block}
\title{Delaware Block Example Data}
\format{
A tibble with population data for Delaware blocks from Census 2010
}
\usage{
data("de_block")
}
\description{
This data set contains block level data for Delaware for 2010.
}
\examples{
data(de_block)
}
\concept{data}
