% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{de_block_group}
\alias{de_block_group}
\title{Delaware Block Group Example Data}
\format{
A tibble with population data for Delaware blocks groups from ACS 2019
}
\usage{
data("de_block_group")
}
\description{
This data set contains block group level data for Delaware for 2019 from
the American Community Survey. This is the five year variant (2015-2019).
}
\examples{
data(de_block_group)
}
\concept{data}
