% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vest_crosswalk.R
\name{vest_crosswalk}
\alias{vest_crosswalk}
\title{Download Processed VEST Block Crosswalk}
\usage{
vest_crosswalk(state)
}
\arguments{
\item{state}{character. The state to get the VEST crosswalk for.}
}
\value{
tibble
}
\description{
Provides a friendlier data format for R for working with VEST crosswalks.
Data can be retallied with \code{PL94171::pl_retally()} using this crosswalk.
}
\examples{
de_cw <- vest_crosswalk('DE')
}
\references{
Amos, Brian, 2021, "2020 Census Block Crosswalk Data",
https://doi.org/10.7910/DVN/T9VMJO, Harvard Dataverse, V2
}
